Implement bare minimum changes required for a billion digits of pi with GMP back end.

+++ include/boost/multiprecision/gmp.hpp
@@ -2260,7 +2260,7 @@ public:
    {
       return -(max)();
    }
-   BOOST_STATIC_CONSTEXPR int digits = static_cast<int>((Digits10 * 1000L) / 301L + ((Digits10 * 1000L) % 301L ? 2 : 1));
+   BOOST_STATIC_CONSTEXPR long long digits = static_cast<long long>((Digits10 * 1000LL) / 301L + ((Digits10 * 1000LL) % 301L ? 2 : 1));
    BOOST_STATIC_CONSTEXPR int digits10 = Digits10;
    // Have to allow for a possible extra limb inside the gmp data structure:
    BOOST_STATIC_CONSTEXPR int max_digits10 = Digits10 + 2 + ((GMP_LIMB_BITS * 301L) / 1000L);
@@ -2340,7 +2340,7 @@ public:
    static number_type (min)() { return number_type(); }
    static number_type (max)() { return number_type(); }
    static number_type lowest() { return number_type(); }
-   BOOST_STATIC_CONSTEXPR int digits = 0;
+   BOOST_STATIC_CONSTEXPR long long digits = 0;
    BOOST_STATIC_CONSTEXPR int digits10 = 0;
    BOOST_STATIC_CONSTEXPR int max_digits10 = 0;
    BOOST_STATIC_CONSTEXPR bool is_signed = false;
